+++ opencascade-hs/cpp/hs_types.h
@@ -101,6 +101,8 @@ typedef void HLRBRep_Algo;
 typedef void HLRBRep_HLRToShape;
 typedef int HLRBRep_TypeOfResultingEdge;
 typedef void BOPAlgo_Builder;
+typedef void BOPAlgo_BOP;
+typedef int BOPAlgo_Operation;
 
 #define Handle(X) void
 #define ARRAY_1(X) void