+++ opencascade-hs/src/OpenCascade/BOPAlgo/Types.hs
@@ -2,6 +2,12 @@
 {-# LANGUAGE EmptyDataDecls #-}
 module OpenCascade.BOPAlgo.Types
 ( Builder
+, BOP
 ) where
 
+import qualified OpenCascade.Inheritance as Inheritance
+
 data Builder
+data BOP
+
+instance Inheritance.SubTypeOf Builder BOP